fixed casting of Javascript numbers and payment amount calculation
temporary solution, I will have to revise number types.

As to payment amount calculation, when the fund AA has 0 total shares
and no asset balance, the payment calculation did not work.

+++ src/jsMain/kotlin/ledger/obyte/obytejs/ObyteJs.kt
@@ -41,6 +41,6 @@ external interface GetAaStateVarsRequest {
 }
 
 external interface BalanceResponse {
-    val stable: Long
-    val pending: Long
+    val stable: Double
+    val pending: Double
 }
